+++ ot/backend.py
@@ -1867,9 +1867,7 @@ class JaxBackend(Backend):
         if not isinstance(size, int):
             raise ValueError("size must be an integer")
         if type_as is not None:
-            return jnp.asarray(
-                jax.random.permutation(subkey, size), dtype=type_as.dtype
-            )
+            return jnp.asarray(jax.random.permutation(subkey, size))
         else:
             return jax.random.permutation(subkey, size)
 
@@ -2444,7 +2442,6 @@ class TorchBackend(Backend):
             )
             return torch.randperm(
                 n=size,
-                dtype=type_as.dtype,
                 generator=generator,
                 device=type_as.device,
             )
@@ -2909,7 +2906,7 @@ class CupyBackend(Backend):  # pragma: no cover
             return self.rng_.permutation(size)
         else:
             with cp.cuda.Device(type_as.device):
-                return cp.asarray(self.rng_.permutation(size), dtype=type_as.dtype)
+                return cp.asarray(self.rng_.permutation(size))
 
     def coo_matrix(self, data, rows, cols, shape=None, type_as=None):
         data = self.from_numpy(data)
@@ -3360,7 +3357,7 @@ class TensorflowBackend(Backend):
             )
         else:
             return tf.random.experimental.stateless_shuffle(
-                tf.range(size, dtype=type_as.dtype), seed=local_seed
+                tf.range(size), seed=local_seed
             )
 
     def _convert_to_index_for_coo(self, tensor):